Fix the failed test on WPT disabling lazy loading timeout
This CL will fix the test issue reported in crbug.com/1329177. This CL
will remove the function call using requestIdleCallback.
requestIdleCallback is not exactly synced with the idle task in the
renderer process so it was not accurate. Instead, set the longer timeout
duration on the test to ensure the idle task is not queued.

Change-Id: I99193ee47fa605ce1c53720d71135e0c0c660933
Bug: 1329177
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3670373
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1007731}
NOKEYCHECK=True
GitOrigin-RevId: 8f952e261d31b49206281de465fa390b5e6e5a58