Keep track of decoder errors in decoder selector
Based on https://docs.google.com/document/d/15oPtBw-XQ0ee5roSdPVyPMDYRIc2c3xZmIdOlZ9ZBjE/edit#

In this CL, we keep track of the failed decode reason and ensure that
it is recorded as the root cause failure when DecoderStream reports
a PipelineStatus back to WMPI.

video_renderer_impl.cc now attaches the DecoderStatus to the
PipelineStatus, which was previously provided, though itself lacking
any root cause.

This requires getting the root cause into
CompleteDecoderReinitialization, so that it could be forwarded rather
than a generic kDecoderReinitializationFailed error with no cause.
This requires that OnDecoderSelected take a DecoderStatus::Or<Decoder>
instead of a null-on-failure unique ptr, which caused a cascade of
pushing errors around thru various functions.

While this did require the saving of the initial decode error in
DecoderSelector, it also did allow me to not have to store a
|decoder_| state there, so its at least neutral on stateful complexity.

R=liberato

Bug: 1320506
Change-Id: Ib3630337f06364f65bca5f82d9d3f999adb09948
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3652610
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008083}
NOKEYCHECK=True
GitOrigin-RevId: 76e1ff0414f9fee9d53d33c2dd58b2771b03eb29