Avoid null pointer in HTMLPlugInElement::PluginWrapper
This crash was happening because LocalFrame::Client() was returning
nullptr. This patch avoids the code which uses LocalFrame::Client() when
it is a nulllptr.

I was able to reproduce the crash locally with the clusterfuzz repro
script and verified that this fixes the crash. I didn't add a test case
because the clusterfuzz test case is not minimized and requires special
GN args.

Fixed: 1316686
Change-Id: I39ebb59cc1200dce37284b634aa7ad39d43fe305
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3673857
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008353}
NOKEYCHECK=True
GitOrigin-RevId: 0d2c99ba31474a31750c17f178d0ac6ca9546017