Eliminate base::LowerCaseEqualsASCII in /third_party/blink/common
This CL is a behavior-preserving refactor. It replaces calls to
base::LowerCaseEqualsASCII with base::EqualsCaseInsensitiveASCII.

We are deprecating base::LowerCaseEqualsASCII because its interface is a
mild footgun, in that it requires the second parameter to be lowercase.

See the design doc for more details.
https://docs.google.com/document/d/18zWV0zN4oAqMZpond1Ut0QviHUeHCSwCo08mqThwHkQ/edit?usp=sharing

This CL was uploaded by git cl split.

R=haraken@chromium.org

Bug: 1326529
Change-Id: I57806e8bdc7f104a989ab80c5f506de7d65c899b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3657582
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008373}
NOKEYCHECK=True
GitOrigin-RevId: c17b03284aeb50ba58f13176346892b2713d81a1