Fix unpainted selection start/end
Composited selection relies on recording the dimensions of the selection
edges on PaintChunk at paint time. This is currently powered by
the SelectionState as indicated on LayoutObject associated with the
item being painted. This breaks when selection endpoints fall within
LayoutText objects that are not painted. To fix this, we compute a
slightly modified SelectionState that does not include such objects and
uses this state to determine the start/end of selection painting.

Bug: 1318037
Change-Id: I53e5471eef259a305dd11965dc9360370b653329
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3673875
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008615}
NOKEYCHECK=True
GitOrigin-RevId: 67fa20bd2eeee6594ebd717b540e23abebd39ef7