[text-decorations] Prepare |TextDecorationInfo| for decorating box
This patch changes |TextDecorationInfo| to prepare for the
decorating box support, without changing the actual behaviors.

Before this change, when the decoration index is changed by
|SetDecorationIndex|, only |ResolvedThickness| is updated.

This patch changes to update all properties that are computed
from |ComputedStyle| and |Font|. The spec requires they are
computed from the computed style of the [decorating box], not
from the computed style of the target text/box decorations are
painted for.

The actual behavior is not changed, because this patch uses
the target style as the [decorating box] style, as seen in the
TODO comment. This patch makes that, when future patches
change the [decorating box] style to the correct one, all
relevant properties will be updated from the [decorating box].

The "override" capabilities allows callers to use the
specified style or font, instead of the one computed by
|TextDecorationInfo|. With this patch, the NG SVG Text and the
legacy use the "override"s.

If the "override" is used, even when future patches support
computing the correct [decorating box], the [decorating box]
support will be limited. The NG SVG Text and the legacy need
further work to support the [decorating box].

The constructor arguments are adjusted and reordered to
clarify the "override"s.

This patch has no behavior changes.

[decorating box]: https://drafts.csswg.org/css-text-decor-3/#decorating-box

Bug: 1008951
Change-Id: I5da6ee64bf6a5d50d66e3e1f643adaf65d21dbb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3674736
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008681}
NOKEYCHECK=True
GitOrigin-RevId: 6c92cc8410ecfa2d297777f588f941040cbc9e4b