Make RenderFrameHost used for notification permission decision
If RenderFrameHost is available in
PermissionContextBase::GetPermissionStatus, fenced frames are recognized
and proper permission will be set there.
https://source.chromium.org/chromium/chromium/src/+/main:components/permissions/permission_context_base.cc;l=281;drc=f10df0040d17b75ac96491b1c009550e97778bc9;bpv=0;bpt=1

However, existing implementation does make the function use
RenderFrameHost for decision, and a fenced frame can show notification
if it is allowed in the primary main frame now.

This code propagates RenderFrameHost instance to
PermissionContextBase::GetPermissionStatus from where it is created
so that the frame can be recognized as the fenced frame.

Bug: 1271027
Change-Id: I4ed106b9e213e1380eca1b2e3092e3f427888eb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3644598
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008947}
NOKEYCHECK=True
GitOrigin-RevId: c1451af5a7c2f187d4fb8406bf1affc89c72d436