Implement applyConstraints for Transferred MSTs
Adds a queue to record the pending applyConstraints requests that a transferred MediaStreamTrack might receive before it is set up fully. Once the setup is done, all these constraints are applied in order.

This CL also adds a helper function of the same name 'applyConstraints' which takes in ScriptPromiseResolver argument instead of ScriptState. This allows us to store the pair (ScriptPromiseResolver*, MediaTrackConstraints) in the queue.

Bug: 1288839
Change-Id: Iabb3356d40e0357bec135b75d9455b08f6b8b053
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645068
Reviewed-by: Tony Herre <toprice@chromium.org>
Commit-Queue: Palak Agarwal <agpalak@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1009525}
NOKEYCHECK=True
GitOrigin-RevId: 7b9e4c78a0072a8197a8463886f904699556efd5