Add TextControlSetValueSelection::kSetSelectionToStart and use it for HTMLTextAreaElement Cloning
Based on HTML standard, the behavior of element.cloneNode(true)
should be to not clone the selection and clamp selection start/end to 0.
This CL fixes the textarea element, which is clamping to the end.

Fixed: 1329942
Bug: 1330747
Change-Id: I2e28032b0ff02669afbbe9fb7df721fe26ddbeab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3681843
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Di Zhang <dizhangg@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1009721}
NOKEYCHECK=True
GitOrigin-RevId: 93bc43525ce59888d5937b1a010977bfec343772