Add acceleration detection to the gpu benchmarking extension.
This CL adds isAcceleratedCanvasImageSource to the gpu benchmarking
extension.  This will allow tests to confirm that they are covering
the code path that the test was intended to cover.  There are
currently lapses in coverage due to the fact that some test are
not testing what we think they are testing.

This CL targets chrome-based tests, which is why the API was added
to the gpu benchmarking interface.  If a follow-up CL, we may need
to add a similar API to the internals interface, for conten_shell
based tests.

Because the gpu benchmarking extension is external to blink, a new
entry point had to be added to blink's public API.

As a proof of concept, the Pixel_CanvasUsesOverlay and
Pixel_UnacceleratedCanvasUsesOverlay tests were modified to use
this new API.

Bug: 1327672, 1309214
Change-Id: I54960c6339988bf86fe69d195ee6af909dc7f07a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3679793
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010282}
NOKEYCHECK=True
GitOrigin-RevId: 83b30ff8a3548feb06c130b5aee00386e7fc479d