Make a fencedframe has its sandbox flags as OpenerSandboxFlags.
OpenerSandboxFlags was left default for a fencedframe, and it is kNone.
If OpenerSandboxFlags is used, even if the frame is opened by fenced
frame, it may not have the sandbox flags set to fenced frames.

This change enforces having the sandbox flags for fenced frames.

Change-Id: I15fe60bf6c5b75b2a0289aba8c20dc44c8444bf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3667296
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010460}
NOKEYCHECK=True
GitOrigin-RevId: 31325f16048568d97678d0fe8c9c119b5881c28f