Remove COPY from trace event macro usage.
This seems to be causing privacy filtering in chrometto. The reason is that chrometto can't be sure the copied (dynamic) strings don't contain sensitive user information, so it strips them. We believe they won't be stripped if we do not do this copy.

Bug: 1322997
Change-Id: I665ac424afc743e011fd18c95233ff2f1da6d6c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3687351
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Annie Sullivan <sullivan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010545}
NOKEYCHECK=True
GitOrigin-RevId: d3558c21d0a9885647d02f3ce7f4397b199da2d3