[Region Capture] Expand produceCropId() to receive any Element
As per the new spec, produceCropId() now receives any type
of Element. It raises an error if the type is not yet
verified as supported. At the moment, only divs and
iframes are supported, but this will be extended.

Note: The spec actually speaks of CropTarget.fromElement().
That transition will be done in upcoming CLs.

Bug: 1332641
Change-Id: I53e880c5afc2d4bfa6ac520ed054b1f60373c392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3689554
Reviewed-by: Simon Hangl <simonha@google.com>
Commit-Queue: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010890}
NOKEYCHECK=True
GitOrigin-RevId: ed8d7825ae7313ed2e19dc383a4d47c297a0ae07