[text-decorations] Decorating box support in |TextDecorationInfo|
This patch adds support for the decorating box to
|TextDecorationInfo|.

With this change, some functions of |TextDecorationInfo|
return values of the decorating box, instead of the target
box. The order of functions in the header file was adjusted.

For the code to work properly, |DecoratingBoxes| in the
|NGInlinePaintContext| must be set up properly from the tree
before painting decorations. This part is not landed yet.

This patch has no behavior changes because the feature is
under a runtime flag.

Bug: 1008951
Change-Id: Ic8d3b1f4eb7d6a4404e8300868dc888210c68ace
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3687263
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010941}
NOKEYCHECK=True
GitOrigin-RevId: 24b2df501a292a78644d7de994f561c4c541078f