Speculative crash fix for TextFragmentAnchor::FireBeforeMatchEvent
In FireBeforeMatchEvent, I forgot to check to see if |range| actually
has nodes in it before accessing and using the first node in the range.
This patch adds checks to see if the range is empty and if it is still
connected to the document. These same checks exist for the find-in-page
version of the same logic in TextFinder::Scroll.

I feel fairly confident that this speculative fix will stop the crashes.

Fixed: 1330844
Change-Id: I3871f0a76018244dabd1d73c1ebfec30be19e676
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3690158
Auto-Submit: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1011085}
NOKEYCHECK=True
GitOrigin-RevId: 39f57fee245f9d34367c3d2d3cfd9b8e198213e8