Update active aria-modal dialog if detached when getting focused object
On platforms which prune the accessibility tree when a modal ARIA dialog
is shown HandleFocusedUIElementChanged calls UpdateActiveAriaModalDialog
which sets the value of active_aria_modal_dialog_ and then marks the
entire tree dirty if that value changed. Given a newly-focused element
which descends from a newly-unhidden aria-modal dialog, the stored
active dialog may have become detached and replaced due to the tree
update. This breaks the logic we depend on to keep the trees updated and
can prevent the announcement of the newly-shown dialog and any focused
descendant.

AX-Relnotes: Fixed VoiceOver presentation of newly-shown aria-modal
dialogs which were in the DOM but hidden.

Bug: 1311826
Change-Id: I0c3c41de3ed5db68cb090a8f95a4a5c41ae31d5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3655949
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1011386}
NOKEYCHECK=True
GitOrigin-RevId: 8c1d90828b89c87f519b6dcd307af31e18dc39ad