Fix invalidation bug on logical combinations inside :has() #2
Fix the invalidation bug on changing ancestors or ancestor siblings
of :has() anchor element when we have a style rule using logical
combinations containing complex selectors inside ':has()'.

To fix the bug, RuleFeatureSet adds features to invalidation sets
for logical combinations inside :has().

For example, in case of having a style rule '.a:has(:is(.b .c)) {}',
RuleFeatureSet adds features to invalidation sets as if we have
'.b .a {}' so that StyleEngine can invalidate descendant '.a' of
an element if the element have a change on the class value 'b'.

We can support more complex cases with this simple approach.

For '.a:has(~ :is(.b ~ .c .d ~ .e)) ~ .f .g {}', RuleFeatureSet
adds features as if we have these rules:
- '.b ~ .c .a ~ .f .g {}'
- '.d ~ .a ~ .f .g {}'

Still we have invalidation bug on the logical combinations inside
:has() when the mutation happens on the sibling of the :has() anchor
element. (For the above complex case, RuleFeatureSet should add
features as if we have this rule also : '.b ~ .a ~ .f .g {}'. This
will be handled in a separated CL.

Bug: 669058, 1331207
Change-Id: I4f9ded2ec382a6aa72dbdb0a98431ef5818b4a60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3680419
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Byungwoo Lee <blee@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1011413}
NOKEYCHECK=True
GitOrigin-RevId: c0ff783876795c4e48ae82a573f313edd2db1913