Support for repeated table headers and footers in fragmentation.
This adds support for repeated table headers and footers to the NG table
fragmentation engine, and also generic block fragmentation support for
repeated content. The latter will also be needed in order to support
repeated fixed-positioned elements for printing, in the future.

Ideally, we shouldn't really have to clone the fragments generated by
repeatable content, but rather just insert the same fragment into every
parent table fragment, but that's currently not possible, because of
pre-paint, paint, and possibly other things too. Pre-paint and paint
rely on FragmentData objects, which contain the "global-ish" paint
offset for each fragment, so, unless we want to add logic for this on
the pre-paint / paint side, this will have to do for now.

Add an NGFragmentRepeater class, which takes care of the fragment
cloning. It will create a deep clone of the fragment subtree, and update
the break tokens, to tie everything together (so that it's only the
final fragment that doesn't have a break token), and set correct
sequence numbers. This way pre-paint and paint will work correctly "out
of the box".

This fixes some existing (non-WPT) tests. Wrote a few more (WPT) tests,
to shed some light on some of the issues that we need to handle. I
thought about different ways of implementing this, such as actually
laying out every repeated fragment, rather than cloning them, but that
proved to be difficult, since we also have to repeat monolithic content,
including atomic inlines, and also multicol (in addition, this approach
would introduce a lot of repeated node awareness to the layout engine,
which I think we're better off without). So just cloning the whole
fragment tree turned out to be easier. Keeping this as simple as
possible is important, especially since, with any luck, at some point in
the future, we may be able to just re-use the same fragment for each
repetition.

Bug: 1078927
Change-Id: Ie1c9ce587f875716349e5bf1169c6c84f409d4f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3689493
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1011833}
NOKEYCHECK=True
GitOrigin-RevId: efd8f9c31239adeeacb892591cca33726cfae678