Fix ChildHasExplicitInheritance flag with independent inheritance.
When the independently inherited properties optimization was enabled,
we would not correctly set the ChildHasExplicitInheritance flag on the
parent. This could later cause the optimization to be incorrectly
applied on other elements.

This is the second attempt at fixing this bug; the first one simply
turned off the optimization entirely if a child had explicit
inheritance. However, on a closer examination, this should be unneeded;
there was already a check in place to turn it off if a child had
explicit inheritance _and_ any non-inherited property had changed.
This check should have triggered -- and does if we propagate the flag
properly up to the parent style.

Fixed: 1325340
Change-Id: I13a75b1b528447bbd12d5a1d3ab9e53482cbac0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3695175
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1011910}
NOKEYCHECK=True
GitOrigin-RevId: 800033bd61fcf9d5c672d62dd6f6089dbf5599dd