[CDP] Add font display to CSS.FontFace object
When a CSS.fontsUpdated event is dispatched, a FontFace object  (https://chromedevtools.github.io/devtools-protocol/tot/CSS/#type-FontFace) is contained in the event. This object is built from the properties of a blink::FontFace instance. However not all of the latter
properties are included in the former. An example of this is the `fontDisplay` property, which this CL now includes.

The purpose of this is to enrich the data used by clients that consume this event, like Chrome DevTools.

Bug: none
Change-Id: I0a42b2b2f5e4ebda0a72f6c3182804a018074171
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3693747
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1011912}
NOKEYCHECK=True
GitOrigin-RevId: e6d3a14494b3e9b3014a0c1f903ae64c4f0ed0ef