Invalidate when PendingLayer effect node changes.
This is needed because, when the result of DecompositeEffect changes, it
can change the effect node on the PendingLayer without changing the
effect node on any of the paint chunks.

Both added tests fail reliably without this change and pass with it.

Fixed: 1330438
Change-Id: I1fd5a92dbf582d888dac04751c915c1049f7e659
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3692324
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012030}
NOKEYCHECK=True
GitOrigin-RevId: c81eff93c37a98460e05c41b14c72f7530b45d45