Avoid dummy NG pagination when falling back to legacy pagination.
When we need to fall back to legacy pagination, because there's
something inside we can't handle on the NG side (such as tables), the
legacy fallback machinery won't mark the LayoutNGView for legacy
fallback (we don't allow reattaching the layout view), only all its
descendants.

We used to end up with a fragment tree like this when falling back to
legacy printing:

.:: LayoutNG Physical Fragment Tree ::.
* Box (out-of-flow-positioned block-flow)(self paint) LayoutNGView #document
    Box (page block-flow)
      Box (block-flow-root legacy-layout-root block-flow)(self paint) LayoutBlockFlow HTML

Check this in NGBlockNode::IsPaginatedRoot(), so that we don't enter
dummy NG pagination, in addition to real legacy pagination.

I'm not aware of any actual problems currently caused by this, but I'm
working on a CL that requires the pre-paint code to tell the truth about
NG-ness - which it didn't, because of this.

Bug: 1121942
Change-Id: I6d5c9aef2106d38357c2342ba9a0dbacbf709e01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3695398
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012098}
NOKEYCHECK=True
GitOrigin-RevId: cee75dc4942b1391fa8a0939cb242ed0083175cd