Add a style perftest.
This has been in use for a while, but is now checked in (except the data
files for now). This verifies style performance (and also hooks into layout,
but not generally doing layout itself). This isolates style from paint etc.,
for more stable benchmarking and profiling. Note that this test
depends on external JSON files with stored web pages, which are
not yet checked in. The tests will be skipped if you don't have the
files available.

Change-Id: I18b8ddb2418c85f76d77e942d261136cf34eb338
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3683403
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012122}
NOKEYCHECK=True
GitOrigin-RevId: 91fd8f5aa000e375fd8288c8c46055070f4c6632