[block-in-inline] Fix duplicated items in outline rects
This patch fixes a problem when |NGPhysicalFragment::
AddOutlineRectsForCursor| may add multiple rects for a
fragment. This happens because when it finds a box item,
it calls |AddOutlineRectsForDescendant| that adds all its
descendants, but still moves the cursor to next in the DFS
order.

When deep nesting occurs, it can create the n! number of
rects when n is the nest level of non-culled inline boxes.

This is an old issue, but the block-in-inline made creating
deep trees easier, and making them non-culled in r980432
crrev.com/c/3486321 made it easier to hit.

Bug: 1333283, 1324517
Change-Id: I2ab439219bfa890f34901de6fd8c72d1b44e1b20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696277
Reviewed-by: Kent Tamura <tkent@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012364}
NOKEYCHECK=True
GitOrigin-RevId: 1378e8c7211152bd2ec56d84bbcbe2b649bcd818