fenced-frames: Disable text fragments
It may be ok to allow text fragments in fenced frames but we'll have to
consider whether there's any way to leak scroll offsets across a fenced
frame. This may be easier and more conclusive when their implementations
converge and settle. Until then, it'd be better to block outright so
that text fragments can't be used to leak data across origins.

Bug: 1314858,1334788
Change-Id: Ic8695cd1fcdf4c8769d37466a546588fbc919ea6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3697156
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012529}
NOKEYCHECK=True
GitOrigin-RevId: 58e680f55dd3457302f42b3f83cc705b995fb4f9