fenced-frames: Layout unaffected by browser controls
Browser controls (i.e. the URL bar on Chrome Android) affect the layout
only in the outermost main frame and not inside a fenced frame. Change
the IsMainFrame checks to IsOutermostMainFrame.

Portals _may_ also want to be affected in these cases but it's not clear
yet what the behavior there should be so leaving a TODO there for now.

Bug: 1314858
Change-Id: I84c39e10b972d39ab75f6b1e16e4151fc53ee8c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696839
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012531}
NOKEYCHECK=True
GitOrigin-RevId: 3be485d2df49dc0ad38ded653f3a292f575943a3