Add application bound service API to elevation service.
This new API allows encryption and decryption of data that will
be bound to the caller via a custom Validator.

Current support is for no validation, or path-based validation.

Path based validation should be used when the caller is
located in a 'trusted path' such as C:\Program Files.

Once data has been encrypted with a particular validation level
then it will only decrypt if the validation matches in the
decrypt call.

The service itself does not store any data, but is encrypted
with both the SYSTEM DPAPI context and the DPAPI context of
the caller. This ensures that the data is also user-bound.
As SYSTEM DPAPI context is used, the code resides inside the
existing elevation service that is installed with the browser.

This CL also adds unit tests for new functions.

BUG=1333461

Change-Id: I2868bb719c902174dbc9d8811c7a391018270a5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3691789
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: S. Ganesh <ganesh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012542}
NOKEYCHECK=True
GitOrigin-RevId: ce7145351bd4c2d58e375a6e0365dc521695084a