[blink] Early exit on no-op changes to style and class attributes.
We observed real-world sites (see bug) making repetitive no-op changes
to the style or class attribute of an element, triggering a style recalc
every time. This CL adds a feature to exit early when that happens.
We'll run an experiment to assess the impact on CPU usage.

Most of the complexity in this CL comes from the fact that the early
exit cannot happen when the class attribute is set as part of moving
an element between documents with different "quirks mode". In this case,
the attribute's value must be re-evaluated using the new "quirks mode".

Bug: 1332843
Change-Id: Ifa8787be4e4df4bc6f72a67ec52dc84cb0c4b210
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3689592
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012605}
NOKEYCHECK=True
GitOrigin-RevId: bed5f197d6afed92adfcba02517ddba3b87e0478