Make MediaQuerySet immutable
This CL prepares for implementing a media query cache to address a
memory regression after moving media query classes to Oilpan.

CSSOM (MediaList) needs to mutate the MediaQuerySet, but we can
implement this by replacing the entire MediaQuerySet object instead.
In order to do this, however, we need to replace the object at the
"source", i.e. in the StyleRule or CSSStyleSheet. Therefore we
introduce MediaQuerySetOwner, which MediaList uses to read/write
MediaQuerySets from/to.

This extra indirection removes the need for "reattachment" when
the StyleRule[Base] changes.

Bug: 1312000, 1331260
Change-Id: I4cf3798fa3a206fbf93f028fd112267ebb591059
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3689489
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012620}
NOKEYCHECK=True
GitOrigin-RevId: 5d2596cdd63b2af90b06a1aa5f4cf6973fcba9bb