[Scroll Unif.] Include main thread repaint time in event latency.
For GestureScrollUpdate handled on the compositor thread, we were basing
latency metrics on the partial update presentation time.  But if the
scroll requires a main thread repaint, we should base its latency on the
presentation time of the frame that includes the main thread's updates.

See go/su-latency for more details on this change.

Bug: 1314720
Change-Id: If9b773ba03ad40be79f1756e823d84f9491ced15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3687996
Reviewed-by: Mohsen Izadi <mohsen@chromium.org>
Commit-Queue: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012679}
NOKEYCHECK=True
GitOrigin-RevId: 134c6e21b8247f9f8b22e244bceaff27499effc3