[GridNG] Pass missing subgrid span size to ResolveGridPositionsFromStyle
`GridPositionsResolver` takes an optional `subgrid_span_size` parameter
that is set by default to `kNotFound`. This value was correctly passed
through when determining the explicit grid column/row count of a
subgrid, but it was missing when resolving the actual item positions.

Bug: 618969
Change-Id: Id3aff5f538f49f69f1fe0003abb31eaaf8cf9024
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3697372
Commit-Queue: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1012729}
NOKEYCHECK=True
GitOrigin-RevId: b13d3fe7b3c47a56354ef54b221008afa754412e