[GridNG] Fix behavior of alternating between standalone/subgridded axes
Previously, alternating between standalone and subgridded axes would
produce a bug where a subgridded item contributed to the sizing of a
given axis but was not subgridded in such dimension.

Introducing several flags to determine for which axes a subgrid should
keep delegating its subgridded items, as well as to know if any grid
item should be considered by the sizing algorithm.

Bug: 618969
Change-Id: I6b45b4b1a6b82495806980735d3cc5ee07d76053
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3693997
Commit-Queue: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012745}
NOKEYCHECK=True
GitOrigin-RevId: ce3f0517d4c0197b6bb0cb14dbbbc3af68557391