Make NamedPropertyQuery match NamedPropertyEnumerator in DOMMimeTypeArr
In [1], I hard-coded the plugins list to a fixed array, when not running
in NaCl/PNaCl mode. I added checks in most places for the flag that
controls this behavior, but missed one place, in
DOMMimeTypeArray::NamedPropertyQuery. This CL makes that function
behave like the rest, and query the fixed array when appropriate.

This exposed the fact that the PPAPITest suites (both in content and
chrome) weren't setting the 'AllowNonEmptyNavigatorPlugins' setting, as
they were missed in the CL [2] that added that setting. That is fixed
here also.

Thanks to tvanderlippe@ and crrev.com/c/3303674 for doing most of
the work here.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3017890/
[2] https://chromium-review.googlesource.com/c/chromium/src/+/2660143

Fixed: 1262066
Change-Id: I1b1de47a74696457ff2f134b4befb894c4e4329a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3692096
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: Derek Schuff <dschuff@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012780}
NOKEYCHECK=True
GitOrigin-RevId: aa76c688c8b37ddb0f74603a4b583928e97c4ca2