Fix intrinsic size of input when spin-box width is a percent-value
In the past, even when the spin-box was percent-value, the intrinsic
size of the input was calculated by adding the value as in the case of
fixed-value.

In this patch, When it is percent-value, the following calculation is
performed.

result = result + result * p / (100 -p)
(where p is a percent value (0 <= p < 100))

Bug: 1138700
Change-Id: I76a5db1806628b71c02cf31305bb92bc1d96b50a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3627410
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012850}
NOKEYCHECK=True
GitOrigin-RevId: 0c0e20acbce5c6d71d2fb16898d5e3a6de4c3692