[GridNG] Account for margin/border/padding in subgrid layout
Introducing extra margin data to `NGGridLayoutTrackCollection` to
correctly account for the margin, border, scrollbar gutter, and padding
of a subgrid when placing its grid items. This is necessary because the
grid geometry from the parent grid is inherited as it is and needs to be
translated to the relative border box of the subgrid.

Bug: 618969
Change-Id: I196fdc3496d8c3fc9d8890ebca776e9f8c3c39b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3698258
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Ethan Jimenez <ethavar@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1013105}
NOKEYCHECK=True
GitOrigin-RevId: 979b46b8269f1cc1ed7d8e95f1e78171bd33353b