Avoid double flow thread insertion notification for ex-spanners.
We used to call FlowThreadDescendantWasInserted() for the same
descendant both via RemoveSpannerPlaceholderIfNoLongerValid() and
FlowThreadDescendantStyleDidChange(), causing DCHECK failures.

Skip the call to FlowThreadDescendantStyleDidChange() for spanners (and
ex-spanners). FlowThreadDescendantStyleDidChange() is expected to follow
a previous call to FlowThreadDescendantStyleWillChange(), but
FlowThreadDescendantStyleWillChange() will never be called for
spanners/ex-spanners, while, without this fix
FlowThreadDescendantStyleDidChange() would in fact be invoked for
ex-spanners (a spanner that's about to become regular column content),
which would trigger another FlowThreadDescendantWasInserted() in some
cases.

Bug: 1333166
Change-Id: I5b643df42987e9e8ef3abfa27c6d72535b66f091
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3695343
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1013375}
NOKEYCHECK=True
GitOrigin-RevId: 7916ce5ea87b2cbbccc774024d8cb588648ca518