ServiceWorker: Remove experimental FetchEvent worker timing
FetchEvent worker timing is the experimental feature behind the runtime
flag. The implementation of the feature has been suspended since 2019
November, and we don't have a plan to complete it. For code simplicity,
this CL removes the entire implementation.

This removal should be safe as the feature has never been shipped on any
browsers and its spec is still in the editor's explainer stage[1] (not
merged into the service worker spec and the resource timing spec).

[1] https://github.com/wanderview/fetchevent-worker-timing/blob/master/explainer.md

Bug: 900700
Change-Id: I781ccece189e80bfb329073ed952dc1bffd1d1b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3697967
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1013407}
NOKEYCHECK=True
GitOrigin-RevId: 777a6a936a21e7ab50796bf9eb36df4e07b4e301