Move ClientHintsPreferences processing out of HTMLPreloadScanner
The ClientHintsPreferences logic in HTMLPreloadScanner has to be done on
the main thread. To prepare for experimenting with an off-main-thread
preload scanner, this changes HTMLPreloadScanner to instead collect a
list of Accept-CH meta values which are then processed by
HTMLDocumentParser. This has the added benefit of unifying all Accept-CH
processing logic in HTMLMetaElement::ProcessMetaAcceptCH().

Bug: 1329535
Change-Id: I1e9c72b3cd108f921668dc0a64920e51587a8392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3680785
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1013535}
NOKEYCHECK=True
GitOrigin-RevId: 071b5f31181a8242c67776334f87efdb0f7c35f0