[anchor-position] Compute position-fallback styles
The styles declared in @position-fallback do not affect the computed
style of any element, but still need to be applied to the base style
and used during layout.

Following this idea, this patch computes the position fallback styles
and stores them in the data cache of the base computed style.

Bug: 1309178
Change-Id: Iae5bda5a3acf848e9d8c1f5199e16831d8686649
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696572
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1013673}
NOKEYCHECK=True
GitOrigin-RevId: 52b16b2b54680fa6d13b90c41cc476e853173f4c