[Blink] Rename HasEditableStyle/HasRichlyEditableStyle
This change is part of a refactor that aims at unifying the logic used
by Blink to determine if a node is richly editable and the one used by
accessibility code.

We simply rename the function HasEditableStyle to IsEditable and
HasRichlyEditableStyle to IsRichlyEditable. These functions will, in an
upcoming CL, do more than simply look at the computed style to
determine if a node can be editable/richly editable, thus the renaming.

Bug: 1331359
Change-Id: I836a10c23eafe70c5b5f52933b5bbfd9b122121f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696134
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1013681}
NOKEYCHECK=True
GitOrigin-RevId: a83dcbef0c5dfcedfed830b23b6c3aeeca4d7822