Filter DelegatedInk metadatas with same timestamp
In order to prevent unnecessary work in the GPU process and extra calls
to the windows StartNewTrail method, filter out delegated ink metadatas
in DelegatedInkTrailPresenter::updateInkTrailStartPoint if the
timestamp of the event is the same as its predecessor.

Duplicate timestamps can be the result of the web app calling updateInkTrailStartPoint repeatedly with the same event. This results
in the same metadata being sent to the DelegatedInkPointRenderer, which
causes a drop in points being rendered until a new metadata is sent.

Change-Id: I14de7b3c0aee5c9e71f3665763dc697475437947
Bug: 1335238
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3698883
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014102}
NOKEYCHECK=True
GitOrigin-RevId: 60d445904cd0f83ee1ea099ebc45cc86cad86919