[FilesTrash] Send outputs back on the TrashIOTask event
When the TrashIOTask completes it emits an event, in the current JS
implementation of trash completion creates a toast with an "Undo"
button. This button requires the Entry's available to perform the
restore. To achieve this, the Entry generation needs to change to ensure
the UI can operate on them AND the output FileSystemURLs need to be
converted to Entry's to traverse over the JS bindings.

Bug: b:231830250
Test: unit_tests --gtest_filter=*RestoreIOTask*
Test: unit_tests --gtest_filter=*TrashIOTask*
Test: Manually run startIOTask and ensure Entry's are available
Change-Id: If538943b6f108ead36ab87f09948736f1a6df89d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3699270
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014326}
NOKEYCHECK=True
GitOrigin-RevId: ab3822a23c30ab2b446be248f7502d64a0b37165