Added command and event for a Top Layer support
This CL adds a new command that returns ids of top layer elements as well as an event that is triggered when elements are added or removed from the top layer. The event and command are used to build a new tooling in the Elements panel that provides support for elements in the top layer.

Bug: 1313690
Change-Id: I40d678f4502ee0343638e280b7cdab983a7b74c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3660404
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014447}
NOKEYCHECK=True
GitOrigin-RevId: 24c9c7b4d4f0e613fa2bc3d006fe9699e7ff3a3b