Modified text decoration thickness test to pass on all platforms
The lowercase 'g' descends through the underline,
which is rendered with dilation
(gap between the underline and the descending g).

Due to different anti-aliasing settings on different test platforms,
the font metrics are calculated slightly differently,
causing a 1 px difference between the from-font in the reference and
the manually set length in the test on at least one platform,
depending on the set font size.

To avoid dilation issues, the proposed work around is to change the
lowercase 'g' in the test text to uppercase 'G'.

Fixed: 1068947

Change-Id: Id2be1fb3e56f9faab8a9698522b21cc42b57f5d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3689109
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Traian Captan <tcaptan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014580}
NOKEYCHECK=True
GitOrigin-RevId: 455f905078b4b828519d63642917205d2e150964