Shared Storage: Change `selectUrl()` param `urls` to include metadata.
We update the `urls` param in `SharedStorage::selectURL()` from a
vector of URLs to a vector of dictionaries, with a required entry in
each dictionary being a candidate `url` and an optional
`reporting_metadata` entry consisting of an array of dictionaries, each
containing a pair `report_event` and `report_url`, in order to
enable event-level reporting.

In this CL, we store the metadata for later use. In a follow-up CL, we
will make this metadata available to be called via
`fence.reportEvent()`.

Bug: 1326692,1218540
Change-Id: Ic2ecd0edca11e08cf6681fa5bcec2a318cf28981
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3650846
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014769}
NOKEYCHECK=True
GitOrigin-RevId: a4915e14d066b2447988d13945eca936f28a3a51