[v8][wasm] Register WasmAsyncResolvePromiseCallback
WebAssembly.compile() returns a Promise that resolves when compilation
on VM-worker threads finishes. So far this Promise got resolved through
the V8-internal API. The V8-internal API, however, does not take care of
microtasks, as this is in the responsibility of the embedder.

The CL https://chromium-review.googlesource.com/c/v8/v8/+/3695584 adds
a callback to the V8 API that allows the embedder to take control of
resolving the Promise and taking care of microtasks. This CL provides
this callback for Chrome. The CL
https://chromium-review.googlesource.com/c/v8/v8/+/3695589 will then
use the new API from the V8 side.

R=yukishiino@chromium.org

Bug: v8:12953
Change-Id: I12b2b0ed1496f7fb72b6f174501d5bcc9dd021b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3697149
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Auto-Submit: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014804}
NOKEYCHECK=True
GitOrigin-RevId: b143f0576a4a6004170d4399637cec11fde62b21