Add error handling for backgroundFetch in fenced frame's service worker
This is a follow-up CL for crrev.com/c/3688774

This CL adds an error handling for backgroundFetch APIs called in the
fenced frame's service worker. Like the APIs exposed on the window
context, backgroundFetch.fetch,get,getIds should be blocked in the
worker context if it's under the fenced frame tree.

Not only the WPT tests, but this CL adds a browser test to ensure the
frame type in the service worker is saved to the storage, and check if
it's used in the API error handling when the service worker is stopped
and activated again.

Change-Id: Ic0ebdd8f36b0e939b2da1eec24af095be0e7e4db
Bug: 1276419
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3697316
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014827}
NOKEYCHECK=True
GitOrigin-RevId: 250f7f179d876094da69b77a77bb26b5cc34d34e