[beacon-api] Add Origin Trial support
The plan is to run an experiment to enable the PendingBeacon API in M104 to test its stability & reliability only for some origins. When a page is opted in to the Origin Trial and its User Agent is also in the
experiment, the API should be available.

Bug: 1293679
Change-Id: I34e4037c49e8521f06eb4bc74dad31f3f91fcb2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3709434
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Ming-Ying Chung <mych@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014884}
NOKEYCHECK=True
GitOrigin-RevId: 4e275d8b6c9c5b718a93fd5b1dae723e3c06f5d9