Add a metric to count duplicated image decoding
This patch adds a metric which can detect if an image is decoded
multiple times in one webpage. It could be checked by the number of the sample with a value of 2 in chrome://histograms.

The metric can evaluate duplicated image decoding in real world,
which can help us to design/refine image cache optimization[1].

Bug: 1324041

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3637145

Change-Id: Iad87781ae316ca807b3110d4dee47bb8c286137c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3666558
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014893}
NOKEYCHECK=True
GitOrigin-RevId: 74e4daab0bc1f6845be4c59b5444e73b5dbf7df5