Insecure Content not allowed in Fenced Frame renderer
Add DCHECK to make sure AllowRunningInsecureContent is set to
false for a Fenced Frame root in the renderer. Since we have to guarantee that Insecure Content setting is not applied in a Fenced Frame, force  AllowRunningInsecureContent setting as false in the renderer.

Bug: 1324264
Change-Id: I56fb7aa910101637ef3e88370ab7cf43295448d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3690381
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1015061}
NOKEYCHECK=True
GitOrigin-RevId: 8cdb8c307a0ecd5ef158caa002ca01b7cd4c1705