Flush buffered attribution triggers as soon as possible
Attribution triggers are buffered on the browser side as long as there
is a connected receiver that may yet register a source, up to a max of
5 seconds (crrev.com/c/3547458). If the last such receiver disconnects before the 5-second limit, any buffered triggers are flushed.

However, the AttributionSrcLoader::ResourceClient does not explicitly
reset its remote, instead relying on it to be closed automatically
during the class's destruction, but as a GC'd type, it may not be
destroyed promptly or ever, effectively guaranteeing that triggers will
always be buffered for the full 5 seconds if any source-eligible data
host has connected first.

This change resets the remote as soon as the ResourceClient knows that
no further source registrations are possible, once its NotifyFinished
method has been called.

Bug: 1336797
Change-Id: I944c8d4ce15c923562a148fbcac9bbbbae5ffa8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3708476
Reviewed-by: John Delaney <johnidel@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Quick-Run: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1015177}
NOKEYCHECK=True
GitOrigin-RevId: 1c25267c642ccc494f8207bdcd0558fcff0bae32