Mark selection bounds as hidden when not painted
After CompositeAfterPaint, composited selection has a dependency on the
selection bounds (start and end) being recorded at paint time. These are
then propagated to the compositor as relative to the created layer
(as they are recorded to be paint chunk-relative).

However, there are situations in which the selection bounds are not
painted, for example if they lie outside of the cull rect of the root
scroller. If we don't record anything to send to the compositor, the
observers in the browser process are not aware of the selection.

Instead, if we detect this case, mark the selection bound as hidden,
which is correctly handled further down the pipeline.

A follow up CL will be needed to fix cases when the handles are grabbed
in the presence of hidden selection bounds. The current code just uses
the existing selection bounds which doesn't take into account content
that is newly hidden, e.g., after a scroll.

R=wangxianzhu@chromium.org

Bug: 1265416
Change-Id: Iabe065cf13fa1d2d2dd52bf0629a607254053cf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3711044
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1015546}
NOKEYCHECK=True
GitOrigin-RevId: e825b77279713c266c40ed7f44d8e1e4b53e9d88