[TableFragmentation] Nested repeated table headers
If a table header was nested inside of another, we would end up
breaking forever. This happened because we would set if the header
ConstraintSpace IsRepeated() based on its parent's ConstraintSpace.
Because the parent's ConstraintSpace had IsRepeated() set to true in
this case, the inner header would never terminate repeating itself,
even when RelayoutAsLastTableBox() was called.

To fix this, don't rely on the parent ConstraintSpace to set this.
Also consider what repeat_mode is set to for the current header.

Once this was fixed, we also ran into an issue when cloning these
nested headers (because they were already repeated). To get this
working, NGFragmentRepeater::GetClonableLayoutResult() needed to be
updated to consider if the repeated break token was the result of
a different fragmentation context than the current header being
cloned.

Bug: 1336683
Change-Id: I13e7af8ce91144e4c3c22eca6dc2a45a0f1a9b80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3710528
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1015548}
NOKEYCHECK=True
GitOrigin-RevId: 1618db124929cf4fb512ee0d9ddacabb7ac0f723