Add feature to use time based parser budget
This experiments with using a time based budget for the parser instead
of token based. The idea is that a time based budget could be a better
fit since tokens can vary widely on how expensive they are to parse,
and can be quite different from site to site.

Bug: 1314493
Change-Id: Ifa29a076e285ab7726c722092c478bf16d343672
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3709257
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1015573}
NOKEYCHECK=True
GitOrigin-RevId: 994ea44211ebd3a983570e7f17a7671d13d98f03