[TableFragmentation] Repeated table header test updates
The spec says that a table section may not be taller than 25% of the
fragmentainer [1], in order to repeat. Update a couple of failing tests
to expand the height of the fragmentainers to ensure they pass in NG
given this spec requirement.

[1] https://www.w3.org/TR/css-tables-3/#repeated-headers

Bug: 1336169
Change-Id: Ic2138cd43b75eea67463e066b20daef4b21c8d90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3710673
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1015575}
NOKEYCHECK=True
GitOrigin-RevId: a5c8ef8eedb6d6c935998db282231fa4dc5482d9